Category & Item Item - [ ] 전체 아이템 테이블 생성 - [ ] 스케줄링을 통한 아이템 테이블 업데이트(수량 및 등록) - [ ] item 관련 기본 템플릿들 작성 Category - [ ] 카테고리 추가
다음 작업내용
- [ ] 캐시에 있는 데이터 정보를 받아오는 방법 찾기
- [ ] 아이템 주문 + 장바구니 구현

+++ src/main/java/com/shopper/walnut/walnut/model/dto/BrandDto.java
@@ -14,7 +14,7 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BrandDto {
-    private int brandId;
+    private long brandId;
     private String brandName;
     private String brandPhone;
     private String brandLoginId;
